update endpoints: 1.a) POST /api/v1/schedulers/tests body [test_uuid: "<tests_uuid>"] 1.b) POST /api/v1/schedulers/services body [service_uuid: "<service_uuid>"] 2. update their runners 3. move controllers.
