+++ src/main/groovy/com/github/h2020_5gtango/vnv/lcm/controller/PackageController.groovy
@@ -32,17 +32,14 @@
  * partner consortium (www.5gtango.eu).
  */
 
-package com.github.h2020_5gtango.vnv.lcm.scheduler
+package com.github.h2020_5gtango.vnv.lcm.controller
 
 import com.github.h2020_5gtango.vnv.lcm.model.NetworkService
 import com.github.h2020_5gtango.vnv.lcm.model.PackageMetadata
-import com.github.h2020_5gtango.vnv.lcm.model.TestSuite
-import com.github.h2020_5gtango.vnv.lcm.restclient.TestCatalogue
+import com.github.h2020_5gtango.vnv.lcm.scheduler.Scheduler
 import io.swagger.annotations.ApiResponse
 import io.swagger.annotations.ApiResponses
 import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.web.bind.annotation.GetMapping
-import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RestController
@@ -50,24 +47,14 @@ import org.springframework.web.bind.annotation.RestController
 import javax.validation.Valid
 
 @RestController
-class TestSuiteController {
+class PackageController {
 
     @Autowired
     Scheduler scheduler
 
-    @Autowired
-    TestCatalogue testCatalogue
-
     @ApiResponses(value = [@ApiResponse(code = 400, message = 'Bad Request')])
-    @PostMapping('/api/v1/schedulers/tests')
-    void onChange(@Valid @RequestBody TestSuite testSuite) {
-        def metadata = new PackageMetadata()
-        metadata.testSuites <<testSuite
+    @PostMapping('/api/v1/schedulers')
+    void onChange(@Valid @RequestBody PackageMetadata metadata) {
         scheduler.scheduleTests(metadata)
-    }
-
-    @GetMapping('/api/v1/schedulers/tests/{testUuid}/services')
-    List<NetworkService> listServicessByTestSuite(@PathVariable('testUuid') String uuid) {
-        testCatalogue.findNssByTestSuiteUuid(uuid)
     }
 }
